requirements.txt
json-stream==2.1.1
python-dotenv==0.21.0
tweepy==4.12.1
